[ADEV-204] FIX campaign create tests
