+++ frontend/src/components/form/inputs/inputs.module.scss
@@ -221,6 +221,10 @@
   &[aria-disabled=true] .switchItem {
     cursor: default;
   }
+
+  ion-note {
+    grid-column: 1 / -1;
+  }
 }
 
 .chips {
