+++ frontend/src/service/campaign/api.ts
@@ -40,6 +40,8 @@ export const CampaignAPI = createApi({
       query: (data) => {
         if (data.deadline?.trim()) data.deadline = new Date(data.deadline).toISOString().split('T')[0];
         else data.deadline = null
+        data.desc = data.desc?.trim() ? data.desc.trim() : null;
+        data.instructions_url = data.instructions_url?.trim() ? data.instructions_url.trim() : null;
         return {
           url: '',
             method: 'POST',
